app/requirements.txt
aiohttp
python-dotenv
openai